requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0
bcrypt==3.2.2
asttokens==2.4.1
requests==2.32.3
aiohttp==3.10.0
# LLM support (optional - install only if using API)
# openai>=1.0.0  # Uncomment for OpenAI API
# anthropic>=0.18.0  # Uncomment for Anthropic API
# ollama>=0.1.0  # Uncomment if using Ollama client library
